Added `IssuerConfigurationProperties` and `IssuerEnvironmentExtensions` for managing issuer metadata and environment configurations. Refactored `IssuerContext` to simplify attestation issuers setup and consolidate credential configuration logic.
