+++ src/main/kotlin/eu/europa/ec/eudi/pidissuer/IssuerContext.kt
@@ -27,7 +27,6 @@ import com.nimbusds.oauth2.sdk.dpop.verifiers.DPoPProtectedResourceRequestVerifi
 import com.nimbusds.oauth2.sdk.dpop.verifiers.InMemoryDPoPSingleUseChecker
 import com.nimbusds.oauth2.sdk.id.Issuer
 import com.nimbusds.oauth2.sdk.util.X509CertificateUtils
-import com.nimbusds.openid.connect.sdk.op.OIDCProviderMetadata
 import eu.europa.ec.eudi.pidissuer.adapter.input.scheduler.CredentialRevocationJob
 import eu.europa.ec.eudi.pidissuer.adapter.input.web.IssuerApi
 import eu.europa.ec.eudi.pidissuer.adapter.input.web.IssuerUi
@@ -35,6 +34,7 @@ import eu.europa.ec.eudi.pidissuer.adapter.input.web.MetaDataApi
 import eu.europa.ec.eudi.pidissuer.adapter.input.web.WalletApi
 import eu.europa.ec.eudi.pidissuer.adapter.input.web.security.*
 import eu.europa.ec.eudi.pidissuer.adapter.out.IssuerSigningKey
+import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.IssueMdoc
 import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.learningcredential.IssueLearningCredential
 import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.mdl.*
 import eu.europa.ec.eudi.pidissuer.adapter.out.attestation.pid.*
@@ -57,10 +57,13 @@ import eu.europa.ec.eudi.pidissuer.adapter.out.webclient.KtorHttpClients
 import eu.europa.ec.eudi.pidissuer.adapter.out.webclient.WebClients
 import eu.europa.ec.eudi.pidissuer.domain.*
 import eu.europa.ec.eudi.pidissuer.port.input.*
+import eu.europa.ec.eudi.pidissuer.port.out.attestation.GetAttestationAttributes
 import eu.europa.ec.eudi.pidissuer.port.out.attestation.asDeferred
 import eu.europa.ec.eudi.pidissuer.port.out.jose.GenerateSignedMetadata
+import eu.europa.ec.eudi.pidissuer.port.out.nonce.VerifyNonce
 import eu.europa.ec.eudi.pidissuer.port.out.persistence.GenerateNotificationId
 import eu.europa.ec.eudi.pidissuer.port.out.persistence.GenerateTransactionId
+import eu.europa.ec.eudi.pidissuer.port.out.persistence.StoreIssuedCredential
 import eu.europa.ec.eudi.pidissuer.port.out.proof.ValidateProof
 import eu.europa.ec.eudi.pidissuer.port.out.status.AllocateStatus
 import eu.europa.ec.eudi.pidissuer.port.out.status.GetStatusListTokenStatus
@@ -69,13 +72,11 @@ import eu.europa.ec.eudi.pidissuer.port.out.trust.IsTrustedKeyAttestationIssuer
 import eu.europa.ec.eudi.sdjwt.HashAlgorithm
 import eu.europa.ec.eudi.sdjwt.vc.Vct
 import io.ktor.http.*
-import kotlinx.coroutines.reactor.awaitSingle
 import kotlinx.coroutines.runBlocking
 import kotlinx.datetime.TimeZone
 import kotlinx.serialization.json.Json
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.BeanRegistrarDsl
-import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.boot.http.codec.CodecCustomizer
 import org.springframework.boot.security.oauth2.server.resource.autoconfigure.OAuth2ResourceServerProperties
 import org.springframework.core.env.Environment
@@ -83,9 +84,7 @@ import org.springframework.core.env.getProperty
 import org.springframework.core.env.getRequiredProperty
 import org.springframework.core.io.DefaultResourceLoader
 import org.springframework.core.io.FileSystemResource
-import org.springframework.core.io.Resource
 import org.springframework.http.HttpStatus
-import org.springframework.http.MediaType
 import org.springframework.http.codec.json.KotlinSerializationJsonDecoder
 import org.springframework.http.codec.json.KotlinSerializationJsonEncoder
 import org.springframework.scheduling.annotation.SchedulingConfigurer
@@ -106,22 +105,17 @@ import org.springframework.security.web.server.authorization.HttpStatusServerAcc
 import org.springframework.security.web.server.authorization.ServerWebExchangeDelegatingServerAccessDeniedHandler
 import org.springframework.util.unit.DataSize
 import org.springframework.web.reactive.function.client.WebClient
-import org.springframework.web.reactive.function.client.bodyToMono
 import org.springframework.web.server.WebFilter
-import org.springframework.web.util.UriComponentsBuilder
 import java.net.URI
 import java.net.URL
 import java.security.KeyStore
-import java.security.cert.X509Certificate
 import java.util.*
 import kotlin.time.Clock
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.days
 import kotlin.time.Duration.Companion.minutes
 import kotlin.time.Duration.Companion.seconds
-import kotlin.time.toJavaDuration
 import kotlin.time.toKotlinDuration
-import java.time.Duration as JavaDuration
 
 val log = LoggerFactory.getLogger(PidIssuerApplication::class.java)
 
@@ -132,11 +126,7 @@ fun beans(
     timeZone: TimeZone,
 ) = BeanRegistrarDsl {
     val issuerPublicUrl = env.readRequiredUrl("issuer.publicUrl", removeTrailingSlash = true)
-    val enableMobileDrivingLicence = env.getProperty("issuer.mdl.enabled", true)
-    val enableMsoMdocPid = env.getProperty<Boolean>("issuer.pid.mso_mdoc.enabled") ?: true
-    val enableSdJwtVcPid = env.getProperty<Boolean>("issuer.pid.sd_jwt_vc.enabled") ?: true
     val credentialsOfferUri = env.getRequiredProperty("issuer.credentialOffer.uri")
-    val enableLearningCredential = env.getProperty<Boolean>("issuer.learningCredential.enabled") ?: true
     val trustValidatorServiceUrl = env.getProperty<String>("issuer.trust.service-url")
 
     val issuerKeystore: KeyStore by lazy {
@@ -275,6 +265,26 @@ fun beans(
         }
     }
 
+    fun accessCertificate(): AccessCertificate {
+        val key =
+            when (env.getProperty<KeyOption>("issuer.access-certificate")) {
+                null, KeyOption.GenerateRandom -> {
+                    log.info("Generating random access certificate key for metadata signing")
+                    ECKeyGenerator(Curve.P_256)
+                        .keyID("issuer-kid-1")
+                        .keyUse(KeyUse.SIGNATURE)
+                        .generate()
+                }
+
+                KeyOption.LoadFromKeystore -> {
+                    log.info("Loading access certificate for metadata signing from keystore")
+                    issuerKeystore.loadJwk(env, "issuer.access-certificate")
+                }
+            }
+
+        return AccessCertificate(key)
+    }
+
     //
     // Nonce encryption key
     //
@@ -301,23 +311,7 @@ fun beans(
     // Signed metadata signing key
     //
     registerBean(lazyInit = true) {
-        val key =
-            when (env.getProperty<KeyOption>("issuer.access-certificate")) {
-                null, KeyOption.GenerateRandom -> {
-                    log.info("Generating random access certificate key for metadata signing")
-                    ECKeyGenerator(Curve.P_256)
-                        .keyID("issuer-kid-1")
-                        .keyUse(KeyUse.SIGNATURE)
-                        .generate()
-                }
-
-                KeyOption.LoadFromKeystore -> {
-                    log.info("Loading access certificate for metadata signing from keystore")
-                    issuerKeystore.loadJwk(env, "issuer.access-certificate")
-                }
-            }
-
-        AccessCertificate(key)
+        accessCertificate()
     }
 
     //
@@ -326,51 +320,12 @@ fun beans(
     registerBean { clock }
     registerBean { timeZone }
 
-    val proxy =
-        env.getProperty("issuer.http.proxy.url")?.let {
-            val url = Url(it)
-            val username = env.getProperty("issuer.http.proxy.username")
-            val password = env.getProperty("issuer.http.proxy.password")
-            HttpProxy(url, username, password)
-        }
+    val proxy = httpProxy(env)
     val webClient = WebClients(proxy, secure = "insecure" !in env.activeProfiles)
     registerBean { webClient }
 
     registerBean {
         KtorHttpClients(proxy, secure = "insecure" !in env.activeProfiles)
-    }
-    registerBean {
-        val keycloakProperties =KeycloakConfigurationProperties(
-            env.getRequiredProperty<URL>("issuer.keycloak.server-url"),
-            env.getRequiredProperty("issuer.keycloak.authentication-realm"),
-            env.getRequiredProperty("issuer.keycloak.client-id"),
-            env.getRequiredProperty("issuer.keycloak.username"),
-            env.getRequiredProperty("issuer.keycloak.password"),
-            env.getRequiredProperty("issuer.keycloak.user-realm"),
-        )
-        GetPidDataFromKeyCloak(
-            issuerCountry = env.getRequiredProperty("issuer.pid.issuingCountry").let(::IsoCountry),
-            issuingJurisdiction = env.getProperty("issuer.pid.issuingJurisdiction"),
-            clock = clock,
-            timeZone = timeZone,
-            webClient = bean(),
-            keyCloak = Url(keycloakProperties.serverUrl.toExternalForm()),
-            administrationClient =
-                AdministrationClient(
-                    realm = Realm(keycloakProperties.authenticationRealm),
-                    client = Credentials(username = keycloakProperties.clientId, password = null),
-                    admin =
-                        Credentials(
-                            username = keycloakProperties.username,
-                            password = keycloakProperties.password,
-                        ),
-                ),
-            users = Realm(keycloakProperties.userRealm),
-        )
-    }
-
-    registerBean {
-        GetMobileDrivingLicenceDataMock()
     }
 
     registerBean { DefaultGenerateQrCode() }
@@ -492,181 +447,61 @@ fun beans(
             IsTrustedKeyAttestationIssuer.usingTrustValidatorService(bean(), URI.create(trustValidatorServiceUrl))
         }
     }
+
     registerBean {
-        val verifyKeyAttestation = VerifyKeyAttestation(isTrustedKeyAttestationIssuer = bean())
-        val validateJwtProofWithKeyAttestation =
-            ValidateJwtProofWithKeyAttestation(issuerPublicUrl, verifyKeyAttestation)
-        val validateAttestationProof = ValidateAttestationProof(verifyKeyAttestation)
-        ValidateProof(
-            validateJwtProofWithKeyAttestation,
-            validateAttestationProof,
-            bean(),
-        )
+        getPidDataFromKeyCloak(clock, timeZone, env, webClient = bean())
     }
     registerBean {
+        val ctx =
+            IssuerFactory.Ctx(
+                clock,
+                timeZone,
+                env,
+                credentialIssuerId = issuerPublicUrl,
+                validateProof = validateProof(issuerPublicUrl, bean(), bean()),
+                storeIssuedCredential = bean(),
+                allocateStatus = bean(),
+                generateNotificationId = bean(),
+            )
         val attestationIssuers =
-            buildList {
-                if (enableMsoMdocPid) {
-                    val duration = env.duration("issuer.pid.mso_mdoc.encoder.duration") ?: 31.days
-                    val jwtProofsSupportedSigningAlgorithms =
-                        env.readNonEmptySet(
-                            "issuer.pid.mso_mdoc.jwtProofs.supportedSigningAlgorithms",
-                            JWSAlgorithm::parse,
-                        )
-                    val pidMsoMdocReusePolicy = this@BeanRegistrarDsl.credentialReusePolicy("issuer.pid.mso_mdoc")
-                    val issueMsoMdocPid =
-                        IssueMsoMdocPid(
-                            clock = clock,
-                            getAttestationAttributes = bean(),
-                            issuerSigningKey = getIssuerSigningKey("issuer.pid.mso_mdoc.signing-key"),
-                            deviceBinding =
-                                DeviceBinding.ts3(
-                                    jwtProofsSupportedSigningAlgorithms,
-                                    PreferredKeyStorageStatusPeriod(duration),
-                                ),
-                            credentialReusePolicy = pidMsoMdocReusePolicy,
-                            validity = duration,
-                            validateProof = bean(),
-                            generateNotificationId =
-                                run {
-                                    val enabled =
-                                        env.getProperty<Boolean>("issuer.pid.mso_mdoc.notifications.enabled")
-                                            ?: true
-                                    if (enabled) bean<GenerateNotificationId>() else null
-                                },
-                            storeIssuedCredential = bean(),
-                            allocateStatus = bean(),
-                        )
-                    add(issueMsoMdocPid)
-                    add(issueMsoMdocPid.asDeferred(bean(), bean(), clock))
-                }
+            context(ctx) {
+                buildList {
+                    if (env.getBoolean("issuer.pid.mso_mdoc.enabled") ?: true) {
+                        val issueMsoMdocPid =
+                            IssuerFactory.pidInMdoc(
+                                issuerSigningKey = getIssuerSigningKey("issuer.pid.mso_mdoc.signing-key"),
+                                getAttestationAttributes = bean(),
+                            )
+                        add(issueMsoMdocPid)
+                        add(issueMsoMdocPid.asDeferred(bean(), bean(), clock))
+                    }
 
-                if (enableSdJwtVcPid) {
-                    val expiresIn = env.duration("issuer.pid.sd_jwt_vc.duration") ?: 31.days
-                    val notUseBefore = env.duration("issuer.pid.sd_jwt_vc.notUseBefore")
-                    val pidSdJwtVcReusePolicy = this@BeanRegistrarDsl.credentialReusePolicy("issuer.pid.sd_jwt_vc")
+                    if (env.getBoolean("issuer.pid.sd_jwt_vc.enabled") ?: true) {
+                        val issueSdJwtVcPid =
+                            IssuerFactory.pidInSdJwtVc(
+                                issuerSigningKey = getIssuerSigningKey("issuer.pid.sd_jwt_vc.signing-key"),
+                                getAttestationAttributes = bean(),
+                            )
+                        add(issueSdJwtVcPid)
+                        add(issueSdJwtVcPid.asDeferred(bean(), bean(), clock))
+                    }
 
-                    val digestsHashAlgorithm =
-                        env.getProperty<HashAlgorithm>(
-                            "issuer.pid.sd_jwt_vc.digests.hashAlgorithm",
-                        ) ?: HashAlgorithm.SHA_256
-                    val jwtProofsSupportedSigningAlgorithms =
-                        env.readNonEmptySet(
-                            "issuer.pid.sd_jwt_vc.jwtProofs.supportedSigningAlgorithms",
-                            JWSAlgorithm::parse,
-                        )
-
-                    val issueSdJwtVcPid =
-                        IssueSdJwtVcPid(
-                            clock = clock,
-                            timeZone = timeZone,
-                            getAttestationAttributes = bean(),
-                            issuerSigningKey = getIssuerSigningKey("issuer.pid.sd_jwt_vc.signing-key"),
-                            credentialIssuerId = issuerPublicUrl,
-                            digestsHashAlgorithm = digestsHashAlgorithm,
-                            deviceBinding =
-                                DeviceBinding.ts3(
-                                    jwtProofsSupportedSigningAlgorithms,
-                                    PreferredKeyStorageStatusPeriod(expiresIn),
-                                ),
-                            credentialReusePolicy = pidSdJwtVcReusePolicy,
-                            validity = expiresIn,
-                            validateProof = bean(),
-                            generateNotificationId =
-                                run {
-                                    val enabled =
-                                        env.getProperty<Boolean>("issuer.pid.sd_jwt_vc.notifications.enabled")
-                                            ?: true
-                                    if (enabled) bean() else null
-                                },
-                            storeIssuedCredential = bean(),
-                            allocateStatus = bean(),
-                            calculateNotUseBefore = notUseBefore?.let { duration -> { iat -> iat + duration } },
-                        )
-
-                    add(issueSdJwtVcPid)
-                    add(issueSdJwtVcPid.asDeferred(bean(), bean(), clock))
-                }
-
-                if (enableMobileDrivingLicence) {
-                    val duration = env.duration("issuer.mdl.mso_mdoc.encoder.duration") ?: 31.days
-                    val jwtProofsSupportedSigningAlgorithms =
-                        env.readNonEmptySet(
-                            "issuer.mdl.jwtProofs.supportedSigningAlgorithms",
-                            JWSAlgorithm::parse,
-                        )
-                    val mdlIssuerReusePolicy = this@BeanRegistrarDsl.credentialReusePolicy("issuer.mdl")
-                    val mdlIssuer =
-                        IssueMobileDrivingLicence(
-                            clock = clock,
-                            getAttestationAttributes = bean(),
-                            issuerSigningKey = getIssuerSigningKey("issuer.mdl.signing-key"),
-                            deviceBinding =
-                                DeviceBinding.ts3(
-                                    jwtProofsSupportedSigningAlgorithms,
-                                    PreferredKeyStorageStatusPeriod(duration),
-                                ),
-                            credentialReusePolicy = mdlIssuerReusePolicy,
-                            validity = duration,
-                            validateProof = bean(),
-                            generateNotificationId =
-                                run {
-                                    val enabled =
-                                        env.getProperty<Boolean>("issuer.mdl.notifications.enabled")
-                                            ?: true
-                                    if (enabled) bean() else null
-                                },
-                            storeIssuedCredential = bean(),
-                            allocateStatus = bean(),
-                        )
-                    add(mdlIssuer)
-                    add(mdlIssuer.asDeferred(bean(), bean(), clock))
-                }
-
-                if (enableLearningCredential) {
-                    val issuerSigningKey = getIssuerSigningKey("issuer.learningCredential.signing-key")
-                    val jwtProofsSupportedSigningAlgorithms =
-                        env.readNonEmptySet(
-                            "issuer.learningCredential.jwtProofs.supportedSigningAlgorithms",
-                            JWSAlgorithm::parse,
-                        )
-                    val learningCredentialReusePolicy =
-                        this@BeanRegistrarDsl.credentialReusePolicy("issuer.learningCredential")
-                    val validity = Duration.parse(env.getProperty("issuer.learningCredential.validity", "P31D"))
-                    val digestHashAlgorithm =
-                        env.getProperty<HashAlgorithm>(
-                            "issuer.learningCredential.sdJwtVc.encoder.digests.hashAlgorithm",
-                        ) ?: HashAlgorithm.SHA_256
-
-                    val notificationsEnabled =
-                        env.getProperty<Boolean>("issuer.learningCredential.notifications.enabled") ?: true
-
-                    val issueLearningCredential =
-                        IssueLearningCredential(
-                            sdJwtVcSerialization = SdJwtVcSerialization.Compact,
-                            clock = clock,
-                            getAttestationAttributes = IssueLearningCredential.randomLearningCredentials(clock, bean()),
-                            issuerSigningKey = issuerSigningKey,
-                            digestsHashAlgorithm = digestHashAlgorithm,
-                            deviceBinding =
-                                DeviceBinding.Required(
-                                    jwtProofsSupportedSigningAlgorithms,
-                                    KeyAttestationRequirement.ts3(
-                                        PreferredKeyStorageStatusPeriod(validity),
-                                    ),
-                                ),
-                            credentialReusePolicy = learningCredentialReusePolicy,
-                            validity = validity,
-                            validateProof = bean(),
-                            generateNotificationId = if (notificationsEnabled) bean() else null,
-                            storeIssuedCredential = bean(),
-                        )
-
-                    add(issueLearningCredential)
-                    add(issueLearningCredential.asDeferred(bean(), bean(), bean()))
-                }
-            }.toNonEmptyListOrNull()
-
+                    if (env.getBoolean("issuer.mdl.enabled", true)) {
+                        val mdlIssuer = IssuerFactory.mdl(getIssuerSigningKey("issuer.mdl.signing-key"))
+                        add(mdlIssuer)
+                        add(mdlIssuer.asDeferred(bean(), bean(), clock))
+                    }
+                    if (env.getBoolean("issuer.learningCredential.enabled") ?: true) {
+                        val issueLearningCredential =
+                            IssuerFactory.learningCredentialInSdJwtVc(
+                                issuerSigningKey = getIssuerSigningKey("issuer.learningCredential.signing-key"),
+                                getPidData = bean(),
+                            )
+                        add(issueLearningCredential)
+                        add(issueLearningCredential.asDeferred(bean(), bean(), bean()))
+                    }
+                }.toNonEmptyListOrNull()
+            }
         checkNotNull(attestationIssuers) { "At least one credential issuer must be configured" }
 
         val preferredClientStatusPeriod =
@@ -757,9 +592,9 @@ fun beans(
 
     registerBean {
         GetProtectedResourceMetadata(
-            bean(),
-            enableBearerTokenAuthentication,
-            beanProvider<DPoPConfigurationProperties>().ifAvailable,
+            credentialIssuerMetadata = bean(),
+            bearerTokenAuthenticationEnabled = enableBearerTokenAuthentication,
+            dPoPConfigurationProperties = beanProvider<DPoPConfigurationProperties>().ifAvailable,
         )
     }
 
@@ -998,326 +833,226 @@ fun beans(
     }
 }
 
-private fun BeanRegistrarDsl.credentialReusePolicy(prefix: String): CredentialReusePolicy {
-    val enabled = env.getProperty<Boolean>("$prefix.reusePolicy.enabled") ?: false
-    if (!enabled) return CredentialReusePolicy.None
+private object IssuerFactory {
+    data class Ctx(
+        val clock: Clock,
+        val timeZone: TimeZone,
+        val env: Environment,
+        val credentialIssuerId: CredentialIssuerId,
+        val validateProof: ValidateProof,
+        val storeIssuedCredential: StoreIssuedCredential,
+        val allocateStatus: AllocateStatus,
+        val generateNotificationId: GenerateNotificationId,
+    )
 
-    val type = env.getProperty("$prefix.reusePolicy.type") ?: "ArfAnnex2"
-    return when (type) {
-        "ArfAnnex2" -> {
-            val options = mutableListOf<EudiReusePolicy>()
-            var index = 0
-            while (true) {
-                val detailsStr = env.getPropertyOrEnvVariable("$prefix.reusePolicy.options[$index].details") ?: break
-                val methods = detailsStr.split(",").map { it.trim() }.mapNotNull { EudiReusePolicyType.fromValue(it) }
-                if (methods.isEmpty()) {
-                    index++
-                    continue
-                }
-                methods.mapTo(options) { method ->
-                    when (method) {
-                        EudiReusePolicyType.OnceOnly -> {
-                            val batchSize =
-                                env.getRequiredPropertyOrEnvVariable<Int>("$prefix.reusePolicy.options[$index].batchSize")
-                            val reissueTriggerUnused =
-                                env.getRequiredPropertyOrEnvVariable<Int>(
-                                    "$prefix.reusePolicy.options[$index].reissueTriggerUnused",
-                                )
-                            EudiReusePolicy.OnceOnly(batchSize, reissueTriggerUnused)
-                        }
-
-                        EudiReusePolicyType.LimitedTime -> {
-                            val reissueTriggerLifetimeLeft =
-                                env
-                                    .getRequiredPropertyOrEnvVariable<Long>(
-                                        "$prefix.reusePolicy.options[$index].reissueTriggerLifetimeLeft",
-                                    ).seconds
-                            EudiReusePolicy.LimitedTime(reissueTriggerLifetimeLeft)
-                        }
-
-                        EudiReusePolicyType.RotatingBatch -> {
-                            val batchSize =
-                                env.getRequiredPropertyOrEnvVariable<Int>("$prefix.reusePolicy.options[$index].batchSize")
-                            val reissueTriggerLifetimeLeft =
-                                env
-                                    .getRequiredPropertyOrEnvVariable<Long>(
-                                        "$prefix.reusePolicy.options[$index].reissueTriggerLifetimeLeft",
-                                    ).seconds
-                            EudiReusePolicy.RotatingBatch(batchSize, reissueTriggerLifetimeLeft)
-                        }
-
-                        EudiReusePolicyType.PerRelyingParty -> {
-                            val batchSize =
-                                env.getRequiredPropertyOrEnvVariable<Int>("$prefix.reusePolicy.options[$index].batchSize")
-                            val reissueTriggerUnused =
-                                env.getRequiredPropertyOrEnvVariable<Int>(
-                                    "$prefix.reusePolicy.options[$index].reissueTriggerUnused",
-                                )
-                            val reissueTriggerLifetimeLeft =
-                                env
-                                    .getRequiredPropertyOrEnvVariable<Long>(
-                                        "$prefix.reusePolicy.options[$index].reissueTriggerLifetimeLeft",
-                                    ).seconds
-                            EudiReusePolicy.PerRelyingParty(
-                                batchSize,
-                                reissueTriggerLifetimeLeft,
-                                reissueTriggerUnused,
-                            )
-                        }
-                    }
-                }
-                index++
-            }
-            if (options.isEmpty())
-                CredentialReusePolicy.None
-            else
-                CredentialReusePolicy.EUDI(
-                    id = CredentialReusePolicy.EUDI.ARF_ANNEX_II_ID,
-                    options = options,
-                )
-        }
-
-        else -> {
-            CredentialReusePolicy.None
-        }
-    }
-}
-
-private fun Environment.credentialResponseEncryption(): CredentialResponseEncryption {
-    val isSupported = getProperty<Boolean>("issuer.credentialResponseEncryption.supported") ?: false
-    return if (!isSupported) {
-        CredentialResponseEncryption.NotSupported
-    } else {
-        val parameters =
-            CredentialResponseEncryptionSupportedParameters(
-                algorithmsSupported =
-                    readNonEmptySet(
-                        "issuer.credentialResponseEncryption.algorithmsSupported",
-                        JWEAlgorithm::parse,
-                    ),
-                methodsSupported =
-                    readNonEmptySet(
-                        "issuer.credentialResponseEncryption.encryptionMethods",
-                        EncryptionMethod::parse,
-                    ),
-                zipAlgorithmsSupported =
-                    readNullableNonEmptySet(
-                        "issuer.credentialResponseEncryption.zipAlgorithmsSupported",
-                    ) { algorithm -> algorithm.takeIf { it.isNotBlank() }?.let { CompressionAlgorithm(it) } },
+    context(ctx: Ctx)
+    fun mdl(issuerSigningKey: IssuerSigningKey): IssueMdoc<MobileDrivingLicence> {
+        val validity = ctx.env.duration("issuer.mdl.mso_mdoc.encoder.duration") ?: 31.days
+        val jwtProofsSupportedSigningAlgorithms =
+            ctx.env.readNonEmptySet(
+                "issuer.mdl.jwtProofs.supportedSigningAlgorithms",
+                JWSAlgorithm::parse,
             )
-        val isRequired = getProperty<Boolean>("issuer.credentialResponseEncryption.required") ?: false
-        if (!isRequired) {
-            CredentialResponseEncryption.Optional(parameters)
-        } else {
-            CredentialResponseEncryption.Required(parameters)
-        }
+        val deviceBinding =
+            DeviceBinding.ts3(
+                jwtProofsSupportedSigningAlgorithms,
+                PreferredKeyStorageStatusPeriod(validity),
+            )
+        val mdlIssuerReusePolicy = credentialReusePolicy(ctx.env, "issuer.mdl")
+        val generateNotificationId =
+            ctx.generateNotificationId.takeIf {
+                ctx.env.getProperty<Boolean>("issuer.mdl.notifications.enabled") ?: true
+            }
+        return IssueMobileDrivingLicence(
+            clock = ctx.clock,
+            getAttestationAttributes = GetMobileDrivingLicenceDataMock(),
+            issuerSigningKey = issuerSigningKey,
+            deviceBinding = deviceBinding,
+            credentialReusePolicy = mdlIssuerReusePolicy,
+            validity = validity,
+            validateProof = ctx.validateProof,
+            generateNotificationId = generateNotificationId,
+            storeIssuedCredential = ctx.storeIssuedCredential,
+            allocateStatus = ctx.allocateStatus,
+        )
+    }
+
+    context(ctx: Ctx)
+    fun pidInMdoc(
+        issuerSigningKey: IssuerSigningKey,
+        getAttestationAttributes: GetAttestationAttributes<PidAttributes>,
+    ): IssueMdoc<PidAttributes> {
+        val duration = ctx.env.duration("issuer.pid.mso_mdoc.encoder.duration") ?: 31.days
+        val jwtProofsSupportedSigningAlgorithms =
+            ctx.env.readNonEmptySet(
+                "issuer.pid.mso_mdoc.jwtProofs.supportedSigningAlgorithms",
+                JWSAlgorithm::parse,
+            )
+        val pidMsoMdocReusePolicy = credentialReusePolicy(ctx.env, "issuer.pid.mso_mdoc")
+
+        return IssueMsoMdocPid(
+            clock = ctx.clock,
+            getAttestationAttributes = getAttestationAttributes,
+            issuerSigningKey = issuerSigningKey,
+            deviceBinding =
+                DeviceBinding.ts3(
+                    jwtProofsSupportedSigningAlgorithms,
+                    PreferredKeyStorageStatusPeriod(duration),
+                ),
+            credentialReusePolicy = pidMsoMdocReusePolicy,
+            validity = duration,
+            validateProof = ctx.validateProof,
+            generateNotificationId =
+                ctx.generateNotificationId.takeIf {
+                    ctx.env.getProperty<Boolean>("issuer.pid.mso_mdoc.notifications.enabled") ?: true
+                },
+            storeIssuedCredential = ctx.storeIssuedCredential,
+            allocateStatus = ctx.allocateStatus,
+        )
+    }
+
+    context(ctx: Ctx)
+    fun pidInSdJwtVc(
+        issuerSigningKey: IssuerSigningKey,
+        getAttestationAttributes: GetAttestationAttributes<PidAttributes>,
+    ): IssueSdJwtVcPid {
+        val expiresIn = ctx.env.duration("issuer.pid.sd_jwt_vc.duration") ?: 31.days
+        val notUseBefore = ctx.env.duration("issuer.pid.sd_jwt_vc.notUseBefore")
+        val pidSdJwtVcReusePolicy = credentialReusePolicy(ctx.env, "issuer.pid.sd_jwt_vc")
+
+        val digestsHashAlgorithm =
+            ctx.env.getProperty<HashAlgorithm>("issuer.pid.sd_jwt_vc.digests.hashAlgorithm") ?: HashAlgorithm.SHA_256
+        val jwtProofsSupportedSigningAlgorithms =
+            ctx.env.readNonEmptySet("issuer.pid.sd_jwt_vc.jwtProofs.supportedSigningAlgorithms", JWSAlgorithm::parse)
+
+        return IssueSdJwtVcPid(
+            clock = ctx.clock,
+            timeZone = ctx.timeZone,
+            getAttestationAttributes = getAttestationAttributes,
+            issuerSigningKey = issuerSigningKey,
+            credentialIssuerId = ctx.credentialIssuerId,
+            digestsHashAlgorithm = digestsHashAlgorithm,
+            deviceBinding =
+                DeviceBinding.ts3(
+                    jwtProofsSupportedSigningAlgorithms,
+                    PreferredKeyStorageStatusPeriod(expiresIn),
+                ),
+            credentialReusePolicy = pidSdJwtVcReusePolicy,
+            validity = expiresIn,
+            validateProof = ctx.validateProof,
+            generateNotificationId =
+                ctx.generateNotificationId.takeIf {
+                    ctx.env.getBoolean("issuer.pid.sd_jwt_vc.notifications.enabled") ?: true
+                },
+            storeIssuedCredential = ctx.storeIssuedCredential,
+            allocateStatus = ctx.allocateStatus,
+            calculateNotUseBefore = notUseBefore?.let { duration -> { iat -> iat + duration } },
+        )
+    }
+
+    context(ctx: Ctx)
+    fun learningCredentialInSdJwtVc(
+        issuerSigningKey: IssuerSigningKey,
+        getPidData: GetAttestationAttributes<PidAttributes>,
+    ): IssueLearningCredential {
+        val jwtProofsSupportedSigningAlgorithms =
+            ctx.env.readNonEmptySet(
+                "issuer.learningCredential.jwtProofs.supportedSigningAlgorithms",
+                JWSAlgorithm::parse,
+            )
+        val learningCredentialReusePolicy =
+            credentialReusePolicy(ctx.env, "issuer.learningCredential")
+        val validity = Duration.parse(ctx.env.getProperty("issuer.learningCredential.validity", "P31D"))
+        val digestHashAlgorithm =
+            ctx.env.getProperty<HashAlgorithm>(
+                "issuer.learningCredential.sdJwtVc.encoder.digests.hashAlgorithm",
+            ) ?: HashAlgorithm.SHA_256
+
+        val notificationsEnabled =
+            ctx.env.getProperty<Boolean>("issuer.learningCredential.notifications.enabled") ?: true
+
+        return IssueLearningCredential(
+            sdJwtVcSerialization = SdJwtVcSerialization.Compact,
+            clock = ctx.clock,
+            getAttestationAttributes =
+                IssueLearningCredential.randomLearningCredentials(
+                    ctx.clock,
+                    getPidData,
+                ),
+            issuerSigningKey = issuerSigningKey,
+            digestsHashAlgorithm = digestHashAlgorithm,
+            deviceBinding =
+                DeviceBinding.Required(
+                    jwtProofsSupportedSigningAlgorithms,
+                    KeyAttestationRequirement.ts3(
+                        PreferredKeyStorageStatusPeriod(validity),
+                    ),
+                ),
+            credentialReusePolicy = learningCredentialReusePolicy,
+            validity = validity,
+            validateProof = ctx.validateProof,
+            generateNotificationId = ctx.generateNotificationId.takeIf { notificationsEnabled },
+            storeIssuedCredential = ctx.storeIssuedCredential,
+        )
     }
 }
 
-private fun Environment.readRequiredUrl(
-    key: String,
-    removeTrailingSlash: Boolean = false,
-): HttpsUrl =
-    getRequiredProperty(key)
-        .let { url ->
-            fun String.normalize() =
-                if (removeTrailingSlash) {
-                    this.removeSuffix("/")
-                } else {
-                    this
-                }
-
-            fun String.toHttpsUrl(): HttpsUrl = HttpsUrl.of(this) ?: HttpsUrl.unsafe(this)
-
-            url.normalize().toHttpsUrl()
-        }
-
-private fun <T> Environment.readNonEmptySet(
-    key: String,
-    f: (String) -> T?,
-): NonEmptySet<T> {
-    val nonEmptySet =
-        getRequiredProperty<MutableSet<String>>(key)
-            .mapNotNull(f)
-            .toNonEmptySetOrNull()
-    return checkNotNull(nonEmptySet) { "Missing or incorrect values values for key `$key`" }
-}
-
-private fun <T> Environment.readNullableNonEmptySet(
-    key: String,
-    f: (String) -> T?,
-): NonEmptySet<T>? =
-    getProperty<MutableSet<String>>(key)
-        ?.mapNotNull(f)
-        ?.toNonEmptySetOrNull()
-
-private fun Environment.duration(key: String): Duration? = getProperty(key)?.let { Duration.parse(it) }?.takeIf { it.isPositive() }
-
-internal fun HttpsUrl.appendPath(path: String): HttpsUrl =
-    HttpsUrl.unsafe(
-        UriComponentsBuilder
-            .fromUriString(externalForm)
-            .path(path)
-            .build()
-            .toUriString(),
+private fun getPidDataFromKeyCloak(
+    clock: Clock,
+    timeZone: TimeZone,
+    env: Environment,
+    webClient: WebClient,
+): GetAttestationAttributes<PidAttributes> {
+    val keycloakProperties =
+        KeycloakConfigurationProperties(
+            env.getRequiredProperty<URL>("issuer.keycloak.server-url"),
+            env.getRequiredProperty("issuer.keycloak.authentication-realm"),
+            env.getRequiredProperty("issuer.keycloak.client-id"),
+            env.getRequiredProperty("issuer.keycloak.username"),
+            env.getRequiredProperty("issuer.keycloak.password"),
+            env.getRequiredProperty("issuer.keycloak.user-realm"),
+        )
+    return GetPidDataFromKeyCloak(
+        issuerCountry = env.getRequiredProperty("issuer.pid.issuingCountry").let(::IsoCountry),
+        issuingJurisdiction = env.getProperty("issuer.pid.issuingJurisdiction"),
+        clock = clock,
+        timeZone = timeZone,
+        webClient = webClient,
+        keyCloak = Url(keycloakProperties.serverUrl.toExternalForm()),
+        administrationClient =
+            AdministrationClient(
+                realm = Realm(keycloakProperties.authenticationRealm),
+                client = Credentials(username = keycloakProperties.clientId, password = null),
+                admin =
+                    Credentials(
+                        username = keycloakProperties.username,
+                        password = keycloakProperties.password,
+                    ),
+            ),
+        users = Realm(keycloakProperties.userRealm),
     )
-
-/**
- * Loads a key pair alongside its associated certificate chain as a JWK.
- *
- * This method expects to find the following properties in the provided [environment].
- * - [prefix].alias -> alias of the key pair to load
- * - [prefix].password -> password of the key pair
- *
- * @receiver the [KeyStore] from which to load the [JWK]
- */
-@Suppress("SameParameterValue")
-private fun KeyStore.loadJwk(
-    environment: Environment,
-    prefix: String,
-): JWK {
-    fun property(property: String): String =
-        when {
-            prefix.isBlank() -> property
-            prefix.endsWith(".") -> "$prefix$property"
-            else -> "$prefix.$property"
-        }
-
-    fun JWK.withCertificateChain(chain: List<X509Certificate>): JWK {
-        require(this.parsedX509CertChain.isNotEmpty()) { "jwk must have a leaf certificate" }
-        require(chain.isNotEmpty()) { "chain cannot be empty" }
-        require(this.parsedX509CertChain.first() == chain.first()) {
-            "leaf certificate of provided chain does not match leaf certificate of jwk"
-        }
-
-        val encodedChain = chain.map { Base64.encode(it.encoded) }
-        return when (this) {
-            is RSAKey -> RSAKey.Builder(this).x509CertChain(encodedChain).build()
-            is ECKey -> ECKey.Builder(this).x509CertChain(encodedChain).build()
-            is OctetKeyPair -> OctetKeyPair.Builder(this).x509CertChain(encodedChain).build()
-            is OctetSequenceKey -> OctetSequenceKey.Builder(this).x509CertChain(encodedChain).build()
-            else -> error("Unexpected JWK type '${this.keyType.value}'/'${this.javaClass}'")
-        }
-    }
-
-    val keyAlias = environment.getRequiredProperty(property("alias"))
-    val keyPassword = environment.getProperty(property("password"))?.takeIf { it.isNotBlank() }
-
-    val jwk = JWK.load(this, keyAlias, keyPassword?.toCharArray())
-    val chain =
-        getCertificateChain(keyAlias)
-            .orEmpty()
-            .map { certificate -> certificate as X509Certificate }
-            .toList()
-
-    return when {
-        chain.isNotEmpty() -> jwk.withCertificateChain(chain)
-        else -> jwk
-    }
 }
 
-/**
- * Indicates whether a random key pair should be generated, or a key pair should be loaded from a keystore.
- */
-private enum class KeyOption {
-    GenerateRandom,
-    LoadFromKeystore,
-}
-
-enum class AccessTokenType {
-    DPoP,
-    Bearer,
-    BearerAndDPoPIfAvailable,
-}
-
-/**
- * Configuration properties for Keycloak.
- */
-data class KeycloakConfigurationProperties(
-    val serverUrl: URL,
-    val authenticationRealm: String,
-    val clientId: String,
-    val username: String,
-    val password: String,
-    val userRealm: String,
-) {
-    init {
-        require(authenticationRealm.isNotBlank()) { "'authenticationRealm' cannot be blank" }
-        require(clientId.isNotBlank()) { "'clientId' cannot be blank" }
-        require(username.isNotBlank()) { "'username' cannot be blank" }
-        require(password.isNotBlank()) { "'password' cannot be blank" }
-        require(userRealm.isNotBlank()) { "'userRealm' cannot be blank" }
-    }
-}
-
-/**
- * Configuration properties for Issuer metadata.
- */
-@ConfigurationProperties("issuer.metadata")
-internal data class IssuerMetadataProperties(
-    val preferredClientStatusPeriod: JavaDuration,
-    val display: List<DisplayProperties> = emptyList(),
-) {
-    data class DisplayProperties(
-        val name: String? = null,
-        val locale: String? = null,
-        val logo: LogoProperties? = null,
-    )
-
-    data class LogoProperties(
-        val uri: URI,
-        val alternativeText: String? = null,
+private fun validateProof(
+    credentialIssuerId: CredentialIssuerId,
+    isTrustedKeyAttestationIssuer: IsTrustedKeyAttestationIssuer,
+    verifyNonce: VerifyNonce,
+): ValidateProof {
+    val verifyKeyAttestation = VerifyKeyAttestation(isTrustedKeyAttestationIssuer = isTrustedKeyAttestationIssuer)
+    val validateJwtProofWithKeyAttestation =
+        ValidateJwtProofWithKeyAttestation(credentialIssuerId, verifyKeyAttestation)
+    val validateAttestationProof = ValidateAttestationProof(verifyKeyAttestation)
+    return ValidateProof(
+        validateJwtProofWithKeyAttestation,
+        validateAttestationProof,
+        verifyNonce,
     )
 }
 
-@ConfigurationProperties("issuer.sd-jwt-vc")
-internal data class SdJwtVcProperties(
-    val typeMetadata: List<TypeMetadataProperties>,
-) {
-    init {
-        val vcts = typeMetadata.map { it.vct }
-        require(vcts.size == vcts.distinct().size)
+private fun httpProxy(env: Environment): HttpProxy? =
+    env.getProperty("issuer.http.proxy.url")?.let {
+        val url = Url(it)
+        val username = env.getProperty("issuer.http.proxy.username")
+        val password = env.getProperty("issuer.http.proxy.password")
+        HttpProxy(url, username, password)
     }
-
-    data class TypeMetadataProperties(
-        val vct: String,
-        val resource: Resource,
-    ) {
-        init {
-            require(vct.isNotBlank()) { "'vct' cannot be blank" }
-            require(resource.exists()) { "'resource' must exist" }
-        }
-    }
-}
-
-private suspend fun WebClient.authorizationServerSupportedDPoPJWSAlgorithms(authorizationServerMetadata: URI): NonEmptySet<JWSAlgorithm>? =
-    Either
-        .catch {
-            val metadata =
-                get()
-                    .uri(authorizationServerMetadata)
-                    .accept(MediaType.APPLICATION_JSON)
-                    .retrieve()
-                    .bodyToMono<String>()
-                    .timeout(5.seconds.toJavaDuration())
-                    .awaitSingle()
-            OIDCProviderMetadata.parse(metadata).dPoPJWSAlgs?.toNonEmptySetOrNull()
-        }.getOrElse {
-            log.warn("Unable to fetch Authorization Server metadata. DPoP support will be disabled.", it)
-            null
-        }
-
-private fun Environment.getPropertyOrEnvVariable(property: String): String? =
-    getProperty(property) ?: getProperty(toEnvironmentVariable(property))
-
-private inline fun <reified T : Any> Environment.getRequiredPropertyOrEnvVariable(property: String): T =
-    getProperty<T>(property) ?: getProperty<T>(toEnvironmentVariable(property))
-        ?: throw IllegalStateException("Property $property or environment variable ${toEnvironmentVariable(property)} not found")
-
-private fun toEnvironmentVariable(property: String): String =
-    property
-        .replace(".", "_")
-        .replace("[", "_")
-        .replace("]", "")
-        .replace("-", "")
-        .uppercase()
